rabbit_amqqueue_common: New module extracted from rabbit_amqqueue
The functions moved from rabbit_amqqueue to rabbit_amqqueue_common are
used in rabbitmq-common and rabbitmq-erlang-client. However, the rest of
rabbit_amqqueue is not needed in rabbitmq-common and causes circular
dependencies with the broker.

Once rabbit-common and rabbitmq-erlang-client switched to
rabbit_amqqueue_common, rabbit_amqqueue will be moved to
rabbitmq-server.

[#118490793]